tramites/requirement.txt
asgiref==3.2.7
Django==3.0.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
Unipath==1.1
